requirements.txt
opencv-python>=4.5.5.62
scikit-learn~=1.6.1
pillow~=10.2.0

torch~=2.7.0
numpy~=2.2.6
scipy~=1.15.3
